# axcom-backend/engineering/api/openapi.yaml
...
openapi: 3.0.3
info:
  title: E-Commerce Engine API
  description: Comprehensive API specification for Axiolon Labs E-Commerce Engine. All responses wrap payloads in a standardized success/error envelope.
  version: 1.0.0
servers:
  - url: /
paths:
...
  /api/notifications/send:
    post:
      summary: Send notification to user
      tags:
        - Notifications
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user_id
                - type
                - message
              properties:
                user_id:
                  type: string
                type:
                  type: string
                message:
                  type: string
      responses:
        '200':
          description: Notification dispatched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationResponseEnvelope'
...
  /api/inventory/transfer:
    post:
      summary: Transfer inventory stock
      tags:
        - Inventory
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Inventory transferred
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponseEnvelope'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MessageResponseEnvelope:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            message:
              type: string
              example: Operation completed successfully.

    # Auth schemas
...
    Notification:
      type: object
      properties:
        id:
          type: string
        user_id:
          type: string
        type:
          type: string
        message:
          type: string
        sent_at:
          type: string
          format: date-time
        status:
          type: string
    NotificationResponseEnvelope:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Notification'